pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "resume-tailor"
version = "0.1.0"
description = "CLI tool to tailor LaTeX resumes to job descriptions using LLMs"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
claude = ["anthropic>=0.40.0"]
openai = ["openai>=1.0.0"]
gemini = ["google-genai>=1.0.0"]
all = ["anthropic>=0.40.0", "openai>=1.0.0", "google-genai>=1.0.0"]

[project.scripts]
resume-tailor = "resume_tailor.cli.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
